Bringing fork up to date

+++ src/Core/Reporting/ReportingWorkflow.cs
@@ -19,7 +19,7 @@ namespace NDifference.Reporting
 
 		public event EventHandler ReportsComplete;
 
-		public void RunReports(Project project, IReportingRepository reportWriters, AnalysisResult results, IProgress<ProgressValue> progressIndicator)
+		public void RunReports(Project project, IReportingRepository reportWriters, AnalysisResult results, IProgress<Progress> progressIndicator)
 		{
 			try
 			{
@@ -39,7 +39,7 @@ namespace NDifference.Reporting
 				{
 					IReportFormat format = writer.SupportedFormats.First();
 
-					progressIndicator.Report(new ProgressValue { Description = "Building File Map" });
+					progressIndicator.Report(new Progress("Building File Map"));
 
 					var builder = FileMapBuilder.Map()
 						.UsingProject(project)
@@ -54,12 +54,9 @@ namespace NDifference.Reporting
 
 					writer.Map = builder.Build();
 
-					if (!String.IsNullOrEmpty(project.Settings.SubFolder))
+					if (!String.IsNullOrEmpty(project.Settings.SubFolder) && Directory.Exists(project.Settings.SubPath))
 					{
-						if (Directory.Exists(project.Settings.SubPath))
-						{
-							Directory.GetFiles(project.Settings.SubPath, "*" + format.Extension).ToList().ForEach(x => File.Delete(x));
-						}
+						Directory.GetFiles(project.Settings.SubPath, "*" + format.Extension).ToList().ForEach(x => File.Delete(x));
 					}
 
 					if (!project.Settings.ConsolidateAssemblyTypes)
@@ -69,7 +66,7 @@ namespace NDifference.Reporting
 						{
 							IReportOutput typeOutput = new FileOutput(Path.Combine(project.Settings.SubPath, typeChange.Name.HtmlSafeTypeName() + format.Extension));
 
-							progressIndicator.Report(new ProgressValue { Description = "Generating report for " + typeChange.Name });
+							progressIndicator.Report(new Progress("Generating report for " + typeChange.Name));
 
 							this.ReportStarting.Fire(this, new FileProgessEventArgs { FileName = typeChange.Name });
 
@@ -85,7 +82,7 @@ namespace NDifference.Reporting
 
 						this.ReportStarting.Fire(this, new FileProgessEventArgs { FileName = dllChange.Name });
 
-						progressIndicator.Report(new ProgressValue { Description = "Generating report for " + dllChange.Name });
+						progressIndicator.Report(new Progress("Generating report for " + dllChange.Name));
 
 						writer.Write(dllChange, dllOutput, format);
 
@@ -107,7 +104,7 @@ namespace NDifference.Reporting
 
 					this.ReportStarting.Fire(this, new FileProgessEventArgs { FileName = project.Settings.IndexName });
 
-					progressIndicator.Report(new ProgressValue { Description = "Generating summary report" });
+					progressIndicator.Report(new Progress("Generating summary report"));
 					
 					writer.Write(results.Summary, output, format);
 
@@ -139,17 +136,29 @@ namespace NDifference.Reporting
 
 							if (assemblyTopic != null)
 							{
-								assemblyTopic.Children.Add(new SiteMapTopic
+								try
+								{
+									string fileName = typeChange.Name.Replace('<', '_').Replace('>', '_').Replace(',', '.');
+
+									string link = Path.Combine(project.Settings.SubPath, fileName + format.Extension);
+
+									assemblyTopic.Children.Add(new SiteMapTopic
 									{
 										Id = typeChange.Identifier,
 										Title = typeChange.Name,
-										Link = Path.Combine(project.Settings.SubPath, typeChange.Name + format.Extension),
+										Link = link,
 										Indent = assemblyTopic.Indent + 1
 									});
+								}
+								catch
+								{
+									// file name is badly mangled? 
+								}
 							}
 						}
 					}
 
+					// sitemap file for sandcastle help file builder menu
 					string siteMapFragment = Path.Combine(project.Settings.OutputFolder, project.Settings.IndexName + ".sitemap");
 
 					File.WriteAllText(siteMapFragment, siteMap.ToString());
